Tests for healthy and broken consumers
